fix ci

+++ Dev/Cpp/Test/Backend/GPUParticle.cpp
@@ -481,7 +481,7 @@ class GpuParticleContext
 
 
 		TrailRenderUniformBufferVS trailRenderUniformBufferVSInitData = {
-			{static_cast<float>(texWidth - 1), static_cast<float>(31 - clz32(texWidth)), 0, 0}, matTo2DArray(viewMatrix), matTo2DArray(projMatrix), { trailOffset, TrailBufferSize }
+			{static_cast<float>(texWidth - 1), static_cast<float>(31 - clz32(texWidth)), 0, 0}, matTo2DArray(viewMatrix), matTo2DArray(projMatrix), { static_cast<float>(trailOffset), static_cast<float>(TrailBufferSize) }
 		};
 		trailRenderUniformBufferVS = graphicsDevice->CreateUniformBuffer(sizeof(TrailRenderUniformBufferVS), &trailRenderUniformBufferVSInitData);
 	}
